Add new setting `vct_header_background_type` to Customizer #309
The goal of this setting is to control the appearance of
the header's background-color on the frontend.

Two options:
* default
* custom

With "default" header (navbar) will always have transparent
background-color.

After choosing "custom" you can set any color.

In case when user already has a value for `vct_header_background`
setting and this value differs from the default one (#ffffff) the
code should act as if `vct_header_background_type` is set to
"custom" for backward compatibility.